Split the biothings.utils.common tests
